+++ cmd/checkfleet/main.go
@@ -23,6 +23,7 @@ import (
 	"github.com/Allan-Nava/checkfleet/internal/checks/httpcheck"
 	"github.com/Allan-Nava/checkfleet/internal/checks/nats"
 	"github.com/Allan-Nava/checkfleet/internal/checks/patroni"
+	"github.com/Allan-Nava/checkfleet/internal/checks/postgres"
 	"github.com/Allan-Nava/checkfleet/internal/checks/stream"
 	"github.com/Allan-Nava/checkfleet/internal/engine"
 	"github.com/Allan-Nava/checkfleet/internal/output"
@@ -51,7 +52,7 @@ func main() {
 
 func usage() {
 	fmt.Fprintln(os.Stderr, `uso:
-  checkfleet check <all|certs|http|nats|haproxy|stream|patroni|consul> --config checkfleet.yml [--output text|markdown|json] [--exit-on-bad]
+  checkfleet check <all|certs|http|nats|haproxy|stream|patroni|consul|postgres> --config checkfleet.yml [--output text|markdown|json] [--exit-on-bad]
   checkfleet version`)
 }
 
@@ -108,6 +109,9 @@ func runCheck(args []string) error {
 		return err
 	}
 	if err := add("consul", func() engine.Check { return consul.New(*cfg.Checks.Consul) }, cfg.Checks.Consul != nil); err != nil {
+		return err
+	}
+	if err := add("postgres", func() engine.Check { return postgres.New(*cfg.Checks.Postgres) }, cfg.Checks.Postgres != nil); err != nil {
 		return err
 	}
 	if len(selected) == 0 {